Bills: Functionality for presenting bills for the household or the individual user as requested. Chores: Automatic image generator for chore

+++ households/Bill.swift
@@ -14,9 +14,9 @@ class Bill: PFObject, PFSubclassing {
     @NSManaged var household: PFObject
     @NSManaged var is_complete: Bool
     @NSManaged var desc: String
-    @NSManaged var total_amount: Int
+    @NSManaged var total_amount: Double
     @NSManaged var date_due: NSDate
-    @NSManaged var remaining_amount: Int
+    @NSManaged var remaining_amount: Double
     // @NSManaged var logo: PFFile
 
     
@@ -31,7 +31,7 @@ class Bill: PFObject, PFSubclassing {
         }
     }
     
-    init(creator: PFUser, household: PFObject, is_complete: Bool, desc: String, total_amount: Int,  date_due: NSDate, remaining_amount: Int) {
+    init(creator: PFUser, household: PFObject, is_complete: Bool, desc: String, total_amount: Double,  date_due: NSDate, remaining_amount: Double) {
         super.init()
         self.creator = creator
         self.household = household
